Kvaser: Add function to sync dongle time to system time
Previously, the clock on the device was synced in the constructor of the
`KvaserBus`. However, on most dongles I have available there's a drift
of the clock onboard the dongle compared to the system time of about 30
milliseconds per hour. While that may not sound as much, if can become
an issue in long-running programs that create the Bus object early and
rely on accurate time stamps.

By extracting the clock syncing code from the constructor into a public
member function a user can sync the clock when needed.